requirements.txt
rich
mailtm
groq
requests
python-dotenv
datetime
aiohttp
beautifulsoup4
wheel
setuptools
Pillow
telethon
python-dotenv
pytz
humanfriendly
faker
python-json-logger
pycountry